Prepare pod/VolAttach in use functions for reuse
